# src/auditoria/dto/auditoria.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class AuditoriaDTO {
  @ApiProperty()
  readonly plan_auditoria_id: string;

  @ApiProperty()
  readonly titulo: string;

  @ApiProperty()
  readonly tipo_evaluacion_id: number;

  @ApiProperty()
  readonly cronograma_id: number[] = [];

  @ApiProperty()
  readonly estado_id: number;

  @ApiProperty()
  readonly no_auditoria: number;

  @ApiProperty()
  readonly vigencia_id: number;

  @ApiProperty()
  readonly consecutivo_OCI: string;

  @ApiProperty()
  readonly consecutivo_IE: string;

  @ApiProperty()
  readonly tipo_id: number;

  @ApiProperty()
  readonly macroproceso: number;

  @ApiProperty()
  readonly macroproceso_id: number;

  @ApiProperty()
  readonly proceso_id: number;

  @ApiProperty()
  readonly dependencia_id: number;

  @ApiProperty()
  readonly lider_id: number;

  @ApiProperty()
  readonly responsable_id: number;

  @ApiProperty()
  readonly fecha_inicio: Date;

  @ApiProperty()
  readonly fecha_fin: Date;

  @ApiProperty()
  readonly objetivo: string;

  @ApiProperty()
  readonly alcance: string;

  @ApiProperty()
  readonly criterio: string;

  @ApiProperty()
  readonly rec_tecnologico: string;

  @ApiProperty()
  readonly rec_humano: string;

  @ApiProperty()
  readonly rec_fisico: string;

  @ApiProperty()
  readonly temas: string;

  @ApiProperty()
  activo: boolean;

  @ApiProperty()
  fecha_creacion: Date;

  @ApiProperty()
  fecha_modificacion: Date;
}