# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/plan-auditoria": {
      "post": {
        "operationId": "PlanAuditoriaController_post",
        "summary": "Crear un nuevo plan de auditoria",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanAuditoriaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El plan de auditoria ha sido creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanAuditoriaDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          }
        },
        "tags": [
          "plan-auditoria"
        ]
      },
      "get": {
        "operationId": "PlanAuditoriaController_getAll",
        "summary": "Obtener todos los planes de auditoria",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": " query-Filter.e.g.col1:v1,col2:v2… ",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "fields - Fields returned. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "required": false,
            "in": "query",
            "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit - Limit the size of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "offset - Start position of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populate",
            "required": false,
            "in": "query",
            "description": "populate - show subqueries. Must be a boolean",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve todos los planes de auditoria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanAuditoriaDTO"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "plan-auditoria"
        ]
      }
    },
    "/plan-auditoria/{id}": {
      "get": {
        "operationId": "PlanAuditoriaController_getById",
        "summary": "Obtener un plan de auditoria por Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el plan de auditoria.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanAuditoriaDTO"
                }
              }
            }
          },
          "404": {
            "description": "Plan de auditoria no encontrado."
          }
        },
        "tags": [
          "plan-auditoria"
        ]
      },
      "put": {
        "operationId": "PlanAuditoriaController_put",
        "summary": "Actualizar un plan de auditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanAuditoriaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "El plan de auditoria ha sido actualizado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanAuditoriaDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Plan de auditoria no encontrado."
          }
        },
        "tags": [
          "plan-auditoria"
        ]
      },
      "delete": {
        "operationId": "PlanAuditoriaController_delete",
        "summary": "Eliminar un plan de auditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El plan de auditoria ha sido eliminado exitosamente."
          },
          "404": {
            "description": "Plan de auditoria no encontrado."
          }
        },
        "tags": [
          "plan-auditoria"
        ]
      }
    },
...
    "/auditoria/auditor/{personaId}": {
      "get": {
        "operationId": "AuditoriaController_getByAuditor",
        "summary": "Obtener auditorías por auditor",
        "parameters": [
          {
            "name": "personaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": " query-Filter.e.g.col1:v1,col2:v2… ",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "fields - Fields returned. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "required": false,
            "in": "query",
            "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit - Limit the size of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "offset - Start position of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populate",
            "required": false,
            "in": "query",
            "description": "populate - show subqueries. Must be a boolean",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve las auditorías del auditor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditoriaDTO"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auditoria"
        ]
      }
    },
    "/auditoria/{id}": {
      "get": {
        "operationId": "AuditoriaController_getById",
        "summary": "Obtener una auditoria por Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve la auditoria.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriaDTO"
                }
              }
            }
          },
          "404": {
            "description": "Auditoria no encontrada."
          }
        },
        "tags": [
          "auditoria"
        ]
      },
      "put": {
        "operationId": "AuditoriaController_put",
        "summary": "Actualizar una auditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditoriaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "La auditoria ha sido actualizada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriaDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Contratista no encontrado."
          }
        },
        "tags": [
          "auditoria"
        ]
      },
      "delete": {
        "operationId": "AuditoriaController_delete",
        "summary": "Eliminar una auditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La auditoria ha sido eliminada exitosamente."
          },
          "404": {
            "description": "Auditoria no encontrada."
          }
        },
        "tags": [
          "auditoria"
        ]
      }
    },
...
    "/auditor/{id}": {
      "get": {
        "operationId": "AuditorController_getById",
        "summary": "Obtener un auditor por Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve la auditor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditorDTO"
                }
              }
            }
          },
          "404": {
            "description": "Auditor no encontrada."
          }
        },
        "tags": [
          "auditor"
        ]
      },
      "put": {
        "operationId": "AuditorController_put",
        "summary": "Actualizar un auditor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditorDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "La auditor ha sido actualizada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditorDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Auditor no encontrado."
          }
        },
        "tags": [
          "auditor"
        ]
      },
      "delete": {
        "operationId": "AuditorController_delete",
        "summary": "Eliminar un auditor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La auditor ha sido eliminada exitosamente."
          },
          "404": {
            "description": "Auditor no encontrada."
          }
        },
        "tags": [
          "auditor"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan AnualAuditoria CRUD",
    "description": "API CRUD para la gestión de las auditoría y planes de mejoramiento",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "PlanAuditoriaDTO": {
        "type": "object",
        "properties": {
          "objetivo": {
            "type": "string"
          },
          "alcance": {
            "type": "string"
          },
          "criterio": {
            "type": "string"
          },
          "recurso": {
            "type": "string"
          },
          "creado_por_id": {
            "type": "number"
          },
          "vigencia_id": {
            "type": "number"
          },
          "aprobado_jefe_dependencia": {
            "type": "boolean"
          },
          "jefe_dependencia_id": {
            "type": "number"
          },
          "aprobado_secretario_tecnico": {
            "type": "boolean"
          },
          "secretario_tecnico_id": {
            "type": "number"
          },
          "auditorias": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activo": {
            "type": "boolean"
          },
          "fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_modificacion": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "objetivo",
          "alcance",
          "criterio",
          "recurso",
          "creado_por_id",
          "vigencia_id",
          "aprobado_jefe_dependencia",
          "jefe_dependencia_id",
          "aprobado_secretario_tecnico",
          "secretario_tecnico_id",
          "auditorias",
          "activo",
          "fecha_creacion",
          "fecha_modificacion"
        ]
      },
      "AuditoriaDTO": {
        "type": "object",
        "properties": {
          "plan_auditoria_id": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "tipo_evaluacion_id": {
            "type": "number"
          },
          "cronograma_id": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estado_id": {
            "type": "number"
          },
          "no_auditoria": {
            "type": "number"
          },
          "vigencia_id": {
            "type": "number"
          },
          "consecutivo_OCI": {
            "type": "string"
          },
          "consecutivo_IE": {
            "type": "string"
          },
          "tipo_id": {
            "type": "number"
          },
          "macroproceso": {
            "type": "number"
          },
          "macroproceso_id": {
            "type": "number"
          },
          "proceso_id": {
            "type": "number"
          },
          "dependencia_id": {
            "type": "number"
          },
          "lider_id": {
            "type": "number"
          },
          "responsable_id": {
            "type": "number"
          },
          "fecha_inicio": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_fin": {
            "format": "date-time",
            "type": "string"
          },
          "objetivo": {
            "type": "string"
          },
          "alcance": {
            "type": "string"
          },
          "criterio": {
            "type": "string"
          },
          "rec_tecnologico": {
            "type": "string"
          },
          "rec_humano": {
            "type": "string"
          },
          "rec_fisico": {
            "type": "string"
          },
          "temas": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_modificacion": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "plan_auditoria_id",
          "titulo",
          "tipo_evaluacion_id",
          "cronograma_id",
          "estado_id",
          "no_auditoria",
          "vigencia_id",
          "consecutivo_OCI",
          "consecutivo_IE",
          "tipo_id",
          "macroproceso",
          "macroproceso_id",
          "proceso_id",
          "dependencia_id",
          "lider_id",
          "responsable_id",
          "fecha_inicio",
          "fecha_fin",
          "objetivo",
          "alcance",
          "criterio",
          "rec_tecnologico",
          "rec_humano",
          "rec_fisico",
          "temas",
          "activo",
          "fecha_creacion",
          "fecha_modificacion"
        ]
      },
...
      "AuditorDTO": {
        "type": "object",
        "properties": {
          "auditoria_id": {
            "type": "string"
          },
          "auditor_id": {
            "type": "number"
          },
          "asignado": {
            "type": "boolean"
          },
          "asignado_por_id": {
            "type": "number"
          },
          "auditor_lider": {
            "type": "boolean"
          },
          "activo": {
            "type": "boolean"
          },
          "fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "fecha_modificacion": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "auditoria_id",
          "auditor_id",
          "asignado",
          "asignado_por_id",
          "auditor_lider",
          "activo",
          "fecha_creacion",
          "fecha_modificacion"
        ]
      },
...
    }
  }
}